MSHARED-1285 use an up-to-date scanner instead the newscanner
Currently it could happen that the scanner misses changed files (because
they are not part of the delta) or copies files even if they have not
changed (e.g. because the output has changes).

This uses now a different approach, instead of only handling the delta
files, we scan all inputs and compare if they are up-to-date with the
output.

Original work from Christoph Läubrich. Modified to accomodate original
full rebuild condition: if filters are modified or output directory is
modified.